% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\docType{data}
\name{aggregation}
\alias{aggregation}
\title{Data set containing 7 clusters of varying shapes and sizes}
\format{
Columns x and y containing the x and y coordinates respectively, and a
column containing the true cluster label
}
\usage{
aggregation
}
\description{
Data set containing 7 clusters of varying shapes and sizes
}
\references{
https://github.com/Kchu/CURE-cluster-python
}
\keyword{datasets}
