% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.R
\name{AIC.gfcm}
\alias{AIC.gfcm}
\title{Akaike information criterion}
\usage{
\method{AIC}{gfcm}(object, ..., k = 2)
}
\arguments{
\item{object}{Object of class cm or gfcm.}

\item{...}{Additionally arguments. Not currently used}

\item{k}{Number to control if either AIC or BIC is to be computed (default is 2 equal to AIC).}
}
\description{
This function computes the AIC for a fitted parametric cure model.
}
