% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.R
\name{BIC.cm}
\alias{BIC.cm}
\title{Bayesian information criterion}
\usage{
\method{BIC}{cm}(object)
}
\arguments{
\item{object}{Object of class cm or gfcm.}
}
\description{
This function computes the BIC for a fitted parametric cure model.
}
