% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.R
\name{BIC.gfcm}
\alias{BIC.gfcm}
\title{Bayesian information criterion}
\usage{
\method{BIC}{gfcm}(object, ...)
}
\arguments{
\item{object}{Object of class cm or gfcm.}

\item{...}{Some methods for this generic function may take additional, optional arguments. At present none do.}
}
\description{
This function computes the BIC for a fitted parametric cure model.
}
