% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrudeProb.R
\name{calc.Crude}
\alias{calc.Crude}
\title{Crude event probabilities}
\usage{
calc.Crude(fit, newdata = NULL, type = "cancer", time = NULL,
  last.point = 100, ci = T, ratetable = survexp.dk, link = "crudeprob")
}
\arguments{
\item{fit}{Fitted model to do predictions from. Possible classes are \code{fmcm}, \code{stpm2}, \code{pstpm2}, \code{CureModel}, and \code{flexsurvreg}.}

\item{newdata}{Data frame from which to compute predictions. If empty, predictions are made on the the data which
the model was fitted on.}

\item{type}{Character indicating which measure is desired. Possible values are \code{cancer} (default),
\code{other}, and \code{othertime} (see details).}

\item{time}{Optional time points at which to compute predictions. If empty, a grid of 100 time points between 0
and \code{last.point} is selected.}

\item{last.point}{Constant at which the bound to tie probability is calculated. Default is 100.}

\item{ci}{Logical indicating whether confidence intervals should be computed.}

\item{ratetable}{Object of class \code{ratetable} to compute the general population survival from. Default is survexp.dk.}

\item{Link}{Character indicating the used link function for computing confidence intervals.}
}
\value{
An object of class \code{crude} containing the crude probability estimates
of each individual in \code{newdata}.
}
\description{
Function for computing crude event probabilties from relative survival models.
}
\details{
The types of crude probabilities is the typical measures in relative survival, namely
a crude disease-specific probability (\code{cancer}), the probability of dying from other causes than the
disease (\code{other}), and the conditional probability of eventually dying from other causes than the
disease given survival until time t (\code{othertime}). The proportion of patients bound to die from the
disease can be computed by using \code{cancer} and choosing a sufficiently large time point.
}
\references{
Eloranta, S., et al. (2014) The application of cure models in the presence of competing risks: a tool
for improved risk communication in population-based cancer patient survival. \emph{Lancet}, 12:86.
}
