% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.MRL.R
\name{calc.MRL}
\alias{calc.MRL}
\title{Mean residual lifetime}
\usage{
calc.MRL(fit, time = NULL, ci = T, tau = 100, newdata = NULL,
  ratetable = survexp.dk)
}
\arguments{
\item{fit}{Fitted model to do predictions from. Possible classes are \code{fmcm}, \code{stpm2}, \code{pstpm2}, \code{CureModel}, and \code{flexsurvreg}.}

\item{time}{Optional time points at which to compute predictions. If empty, a grid of 100 time points between 0
and \code{tau} is selected.}

\item{ci}{Logical indicating whether confidence intervals should be computed}

\item{tau}{The upper limit of the integral. Default is 100.}

\item{newdata}{Data frame from which to compute predictions. If empty, predictions are made on the the data which
the model was fitted on.}

\item{ratetable}{Object of class \code{ratetable} to compute the general population survival from.}
}
\value{
A object of class \code{lol} containing the loss of lifetime estiamtes
of each individual in \code{newdata}.
}
\description{
Function for computing mean (restricted) residual lifetime from an estimated relative survival model
}
