% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.colonDC.R
\docType{data}
\name{colonDC}
\alias{colonDC}
\title{Colon cancer data}
\format{A data frame with 15564 rows and 11 variables}
\usage{
data(colonDC)
}
\description{
A dataset containing the individual baseline and follow-up data on more than 15,000
colon cancer patient. The data is a data cleaned version of the \code{rstpm2::colon} data set.
}
\details{
\itemize{
  \item sex. Either female or male
  \item age. Diagnostic age
  \item stage. Clinical stage at diagnosis (either unknown, localised, regional, or distant)
  \item statusDC. Alive indicator with cause of death
  \item subsite. Anatomical subsite of tumour (either coecum and ascending, transverse,
  descending and sigmoid, or other and NOS)
  \item dx. Date of diagnosis
  \item exit. Date of study exit
  \item status. Alive indicator (0 = alive, 1 = dead)
  \item FU. Follow-up time measured in days
  \item FUyear. Follow-up time measured in years
  \item agedays. Diagnostic age in days.
}
}
\keyword{datasets}
