% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_clus.R
\name{dist_clus}
\alias{dist_clus}
\title{Find distance between two clusters}
\usage{
dist_clus(u, v, rep_pts_df)
}
\arguments{
\item{u}{Cluster label for the first cluster}

\item{v}{Cluster label for the second cluster}

\item{rep_pts_df}{A data frame containing the representative points for all clusters}
}
\value{
The distance between the clusters
}
\description{
Find distance between two clusters
}
\examples{
rep_pts_df <- data.frame(x = iris$Sepal.Length, y = iris$Sepal.Width, clus.lab = 1:length(iris$Sepal.Width));
dist(1, 2,rep_pts_df)

}
