% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihoods.R
\name{extract_general}
\alias{extract_general}
\title{Extract general population hazard}
\usage{
extract_general(time, age, sex, date, data = NULL, ratetable = survexp.dk,
  opposite = F)
}
\arguments{
\item{time}{Either a numeric vector of follow-up times or a character indicating which variable is the follow-up time in the data}

\item{age}{Either a numeric vector of age or a character indicating which variable is the age in the data}

\item{sex}{A numeric vector including the time points for which to make predictions.}

\item{date}{A character denoting the type of prediction. Possible values are \code{relsurv} (default), \code{curerate}, and \code{probcure} (see details)}

\item{data}{The data from which to extract variables from. If \code{time}, \code{age}, \code{sex}, and \code{date} are not characters, leave this empty.}

\item{ratetable}{Object of class \code{ratetable} to extract background hazards from. Defaults to survexp.dk}

\item{opposite}{Logical (default \code{FALSE}) to indicate the order in the the ratetable}
}
\value{
An object of class \code{numeric} of yearly hazards
}
\description{
Yearly general population hazards matched on age, gender, and calendar year is extracted from a ratetable
}
