% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.cif.model.R
\name{fit.cif.model}
\alias{fit.cif.model}
\title{Fit spline-based cumulative incidence model}
\usage{
fit.cif.model(formula, data, cens.code = 0, knots = NULL, n.knots = NULL,
  cure = FALSE, knots.time = NULL, n.knots.time = NULL, covariance = T,
  link = "loglog", verbose = T, constr.optim = F, optim.args = NULL,
  ortho = TRUE, ini.types = c("cure", "flexpara"))
}
\arguments{
\item{formula}{Formula for modelling cause specific cumulative incidence function.
A list of formulas can be provided if different modelling is applied to the causes.
Reponse has to be of the form \code{Surv(time, status)}.}

\item{data}{Data frame in which to interpret the variables names in \code{formula}.}

\item{cens.code}{Numeric. The value of the event indicator which indicates censoring.}

\item{knots}{Knots used for the spline baseline effect of the cumulative incidence.
The knots are provided in a list with each entry denoting the knots corresponding to each cause.}

\item{n.knots}{Number of knots for the baseline splines.
The knots are calculated as the equidistant quantiles of the uncensored cause-specific event-times.
\code{n.knots} are provided in a list with each entry denoting the number of knots corresponding to each cause.
If \code{knots} is supplied, this argument will be ignored.}

\item{cure}{List of logicals. If \code{TRUE}, a cure model is fitted for the cause-specific cumulative incidence,
where cure is assumed after the last selected knot. Typically, cure is only assumed for one cumulative incidence function.
If \code{cure} is not a list, the same option is assumed for all causes.}

\item{knots.time}{A named list containing the knots of each of time-varying covariate effect.}

\item{n.knots.time}{A named list, containing the number of knots for the time-varying covariate effects.
The knots are calculated as the equidistant quantiles of the uncensored event-times.
If \code{knots.time} is supplied, this argument will be ignored.}

\item{covariance}{Logical. If \code{TRUE} (default), the covariance matrix is computed.}

\item{verbose}{Logical. If \code{TRUE} status messages of the function is outputted.}

\item{constr.optim}{Logical. If \code{TRUE} the model is fitted using constraints optimization yielding
a non-negative hazard of the uncured (default is \code{FALSE}).
This option is only implemented for \code{linksu = loglog}.}

\item{optim.args}{List with additional arguments passed to \code{optim}.}

\item{ortho}{Logical. If \code{TRUE} (default), all splines are orthogonalized using a QR-decomposition.}

\item{ini.types}{Character vector denoting the executed schemes for computing initial values.}

\item{type}{A character indicating the type of cure model.
Possible values are \code{mixture} (default) and \code{nmixture}.}

\item{linkpi}{Character giving the link function selected for the cure rate.
Possible values are \code{logit} (default), \code{identity}, \code{loglog}, and \code{probit}.}

\item{linksu}{Character giving the link function selected for the survival of the uncured.
Possible values are \code{loglog} (default), \code{logit}, and \code{probit}.}
}
\value{
An object of class \code{fcm}.
}
