% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jain.R
\docType{data}
\name{jain}
\alias{jain}
\title{Data set containing 2 parabolic shaped clusters}
\format{
Columns x and y containing the x and y coordinates respectively, and a
column containing the true cluster label
}
\usage{
jain
}
\description{
Data set containing 2 parabolic shaped clusters
}
\references{
https://github.com/Kchu/CURE-cluster-python
}
\keyword{datasets}
