% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrtest.cm.R
\name{lrtest.cm}
\alias{lrtest.cm}
\title{Likelihood ratio tests for comparing nested cure models}
\usage{
lrtest.cm(model1, model2)
}
\arguments{
\item{model1}{A model fitted by the fit.cure.model-function}

\item{model2}{Another model fitted by the fit.cure.model-function for comparison.}
}
\value{
An object of class "anova" containing the degrees of freedom and log-likelihoods of the two models,
the difference in degrees of freedom, the likelihood ratio Chi-squared statistic (twice the difference in log-likelihoods)
, and the p-value for the asymptotic likelihood ratio test.
}
\description{
Function for doing likelihood ratio tests for class \code{cm}.
}
\details{
One of the models must be nested within the other model,
and both models must be fitted to the same data set.
}
