% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestCluster.R
\name{nearestCluster}
\alias{nearestCluster}
\title{Find closest cluster}
\usage{
nearestCluster(u, rep_pts_df)
}
\arguments{
\item{u}{Cluster label}

\item{rep_pts_df}{A data frame containing the representative points for all clusters}
}
\value{
A list containing the cluster ID of the closest cluster and the closest distance
}
\description{
Find closest cluster
}
\examples{
rep_pts_df <- data.frame(x = threeClus$x, y = threeClus$y, clus.lab = 1:length(threeClus$x));
nearestCluster(1, rep_pts_df)

}
