% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fcim.R
\name{plot.fcim}
\alias{plot.fcim}
\title{Plot function for Flexible mixture cure model}
\usage{
plot.fcim(fit, newdata = NULL, type = "cuminc", time = NULL,
  cause = NULL, xlim = NULL, ylim = c(0, 1), xlab = "Time",
  ylab = NULL, col = 1, ci = NULL, non.parametric = F,
  col.non.para = 2, include.knots = F, add = F, ...)
}
\arguments{
\item{fit}{Object of class \code{cuRe}.}

\item{newdata}{Data frame from which to compute predictions. If empty, predictions are made on the the data which
the model was fitted on.}

\item{type}{Type of prediction to do. Possible values are \code{relsurv} (default) for the relative survival,
\code{curerate} for the cure rate, \code{ehaz} for the excess hazard, and \code{probcure} for the
conditional probability of being cured.}

\item{time}{Optional time points at which to compute predictions.
This argument is not used if type is \code{curerate}.}

\item{xlim}{Limits of the x-axis}

\item{ylim}{Limits of the y-axis.}

\item{col}{Colour of each line.}

\item{ci}{Logical. If \code{TRUE} (default), confidence intervals are added.}

\item{non.parametric}{Logical. If \code{TRUE}, a non-parametic
estimate (Ederer II method) is added to the plot (default is FALSE).}

\item{col.non.para}{Colour of the non-parametric estimate.}

\item{include.knots}{Logical. If \code{TRUE} vertical lines are added at
each knot of the baseline spline function (default is \code{FALSE}).}

\item{xlabel}{Label of the x-axis. Default is "Time".}

\item{ylabel}{Label of the y-axis. Default is "Relative survival" if \code{type = relsurv},
"Excess hazard" if \code{type = ehaz},
"Conditional probability of cure" if \code{type = probcure},
"Disease-specific survival of the uncured" if \code{type = survuncured}.}
}
\value{
A plot containing the predictions of each observation in \code{newdata}.
}
\description{
Plot function associated with the class \code{fmcm}
}
