% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.lts.R
\name{plot.lts}
\alias{plot.lts}
\title{Plot the long term survival}
\usage{
\method{plot}{lts}(x, ylim = NULL, xlim = NULL, ci = T, col = 1,
        ylab = NULL, xlab = "Time", add = F, \dots)
}
\arguments{
\item{x}{Object of class \code{lts}.}

\item{ylim}{Limit of the y-axis.}

\item{xlim}{Limit of x-axis.}

\item{ci}{Logical. If \code{TRUE} (default), confidence intervals are added to the plot.}

\item{col}{Numeric or character indicating the colours of the curves.}

\item{ylab}{Label to be written on the y-axis. If \code{NULL}, this is based on \code{type}.}

\item{xlab}{Label to be written on the x-axis.}

\item{add}{Logical indicating whether to add to current plot window (default is FALSE).}

\item{...}{Further argument passed to \code{plot} and \code{lines}.}
}
\description{
Plot function for the class \code{lts}.
}
