% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.cm.R
\name{predict.cm}
\alias{predict.cm}
\title{Predict function for cure models}
\usage{
predict.cm(fit, newdata = NULL, type = c("relsurv", "ehaz", "probcure",
  "survuncured"), time = NULL, ci = T, pars = NULL)
}
\arguments{
\item{fit}{Object of class \code{cm} to do predictions from.}

\item{newdata}{Data frame from which to compute predictions. If empty, predictions are made on the data which
the model was fitted on.}

\item{type}{Type of prediction to do. Possible values are \code{relsurv} (default) for the relative survival,
\code{curerate} for the cure rate, \code{ehaz} for the excess hazard, \code{probcure} for the
conditional probability of being cured, and \code{survuncured} for the disease-specific survival of the uncured.}

\item{time}{Optional time points at which to compute predictions.}

\item{ci}{Logical. If \code{TRUE}, confidence intervals are computed.}

\item{pars}{Numerical vector containing the parameters values of the model.}
}
\value{
An object of class \code{matrix} including the predictions.
}
\description{
This function is used to make predictions of the cure models.
}
