% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.CureModel.R
\name{predict.CureModel}
\alias{predict.CureModel}
\title{Predict function for cure models}
\usage{
predict.CureModel(fit, newdata, times, type = "relsurv")
}
\arguments{
\item{fit}{An object of type CureModel.}

\item{newdata}{A dataframe with new data to predict from.}

\item{times}{A numeric vector including the time points for which to make predictions.}

\item{type}{A character denoting the type of prediction. Possible values are \code{relsurv} (default), \code{curerate}, and \code{probcure} (see details)}

\item{ci}{Logical indicating whether to include confidence intervals. Default is \code{TRUE}}
}
\value{
An object of class \code{matrix} including the predictions.
}
\description{
This function is used to make predictions of the cure models.
}
\details{
For type \code{relsurv}, relative survival predictions are made, while for \code{curerate}, \eqn{\pi} and for \code{probcure}\cr
the probability\cr
\deqn{\p{\text{prob}| T > t} = \frac{\pi}{\pi + (1 - \pi) S_u(t)}}-
}
