% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.fcim.R
\name{predict.fcim}
\alias{predict.fcim}
\title{Predict function for Flexible parametric cumulative incidence function}
\usage{
predict.fcim(fit, newdata = NULL, type = "cuminc", time = NULL, ci = T,
  pars = NULL)
}
\arguments{
\item{fit}{Object of class \code{fcim} to do predictions from.}

\item{newdata}{Data frame from which to compute predictions. If empty, predictions are made on the data which
the model was fitted on.}

\item{type}{Type of prediction to do. Possible values are \code{cuminc} (default) for the cumulative incidence,
and \code{subhaz} for the sub-distribution hazard.}

\item{time}{Optional time points at which to compute predictions.
This argument is not used if type is \code{curerate}.}

\item{ci}{Logical. If \code{TRUE}, confidence intervals are computed.}

\item{pars}{Numerical vector containing the parameters values of the model.
In general, this argument can be ignored by the user.}
}
\value{
A list containing the predictions of each individual in \code{newdata}.
}
\description{
Function for doing predictions for class \code{fcim}
}
