% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict2.fcm.R
\name{predict.gfcm}
\alias{predict.gfcm}
\title{Predict function for Flexible mixture cure model}
\usage{
predict.gfcm(object, newdata = NULL, type = c("surv", "curerate",
  "probcure", "survuncured", "hazarduncured", "cumhazuncured", "densityuncured",
  "failuncured", "oddsuncured", "loghazarduncured", "hazard", "density", "fail",
  "loghazard", "odds", "cumhaz"), indi = TRUE, time = NULL,
  var.type = c("ci", "se", "n"), pars = NULL, link = NULL, var = NULL,
  keep.attributes = FALSE, use.gr = FALSE, ...)
}
\arguments{
\item{newdata}{Data frame from which to compute predictions. If empty, predictions are made on the data which
the model was fitted on.}

\item{type}{Type of prediction to do. Possible values are \code{relsurv} (default) for the relative survival,
\code{curerate} for the cure rate, \code{ehaz} for the excess hazard, \code{probcure} for the
conditional probability of being cured, and \code{survuncured} for the disease-specific survival of the uncured.}

\item{time}{Optional time points at which to compute predictions.
This argument is not used if type is \code{curerate}.}

\item{pars}{Numerical vector containing the parameters values of the model.
In general, this argument can be ignored by the user.}

\item{fit}{Object of class \code{gfcm} to do predictions from.}

\item{ci}{Logical. If \code{TRUE}, confidence intervals are computed.}
}
\value{
A list containing the predictions of each individual in \code{newdata}.
}
\description{
Function for doing predictions for class \code{gfcm}
}
