% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.calc.Crude.R
\name{quantile.calc.Crude}
\alias{quantile.calc.Crude}
\title{Compute the time to statistical cure using a probability measure}
\usage{
\method{quantile}{calc.Crude}(fit, q = 0.95, newdata = NULL,
  max.time = 20, expected = NULL, ci = TRUE, rmap,
  ratetable = survexp.dk, last.point = 100, reverse = FALSE)
}
\arguments{
\item{fit}{Fitted model to do predictions from. Possible classes are
\code{fmc}, \code{cm}, \code{stpm2}, and \code{pstpm2}.}

\item{q}{Threshold to estimate statistical cure according to.}

\item{newdata}{Data frame from which to compute predictions. If empty, predictions are made on the the data which
the model was fitted on.}

\item{max.time}{Upper boundary of the interval [0, \code{max.time}] in which to search for solution.}

\item{expected}{Object of class \code{list} containing objects of class \code{survexp},
with the expected survival of each row in newdata. If not specified, the function computes the expected
survival.}

\item{ci}{Logical. If \code{TRUE} (default), confidence intervals are computed.}

\item{rmap}{List to be passed to \code{survexp} from the \code{survival} package.
Detailed documentation on this argument can be found by \code{?survexp}.}

\item{ratetable}{Object of class \code{ratetable} used to compute the general population survival.
Default is \code{survexp.dk}}

\item{last.point}{Constant at which the bound to tie probability is calculated. Default is 100.}

\item{type}{Type of life expectancy measure. Possible values are "ll" for the loss of lifetime
and "mrl" for the mean residual lifetime.}
}
\value{
The estimated cure points.
}
\description{
The following function estimates the time to statistical cure using the
probability of eventually dying from cancer.
}
