% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.calc.LL.R
\name{quantile.calc.LL}
\alias{quantile.calc.LL}
\title{Compute the time to statistical cure using the loss of lifetime function}
\usage{
\method{quantile}{calc.LL}(fit, q = 2, newdata = NULL, max.time = 20,
  ci = TRUE, expected = NULL, rmap = NULL, ratetable = survexp.dk,
  tau = 100, type = "ll")
}
\arguments{
\item{fit}{Fitted model to do predictions from. Possible classes are
\code{fmc}, \code{cm}, \code{stpm2}, and \code{pstpm2}.}

\item{q}{Threshold to estimate statistical cure according to.}

\item{newdata}{Data frame from which to compute predictions. If empty, predictions are made on the the data which
the model was fitted on.}

\item{max.time}{Upper boundary of the interval [0, \code{max.time}] in which to search for solution.}

\item{ci}{Logical. If \code{TRUE} (default), confidence intervals are computed.}

\item{expected}{Object of class \code{list} containing objects of class \code{survexp},
with the expected survival of each row in newdata. If not specified, the function computes the expected
survival.}

\item{rmap}{List to be passed to \code{survexp} from the \code{survival} package.
Detailed documentation on this argument can be found by \code{?survexp}.}

\item{ratetable}{Object of class \code{ratetable} used to compute the general population survival.
Default is \code{survexp.dk}}

\item{tau}{The upper limit of the integral. Default is 100.}

\item{type}{Type of life expectancy measure. Possible values are "ll" for the loss of lifetime
and "mrl" for the mean residual lifetime.}
}
\value{
The estimated cure points.
}
\description{
Compute the time to statistical cure using the loss of lifetime function
}
