% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mixcure.R
\name{summary.CureModel}
\alias{summary.CureModel}
\title{Summary of parametric cure models}
\usage{
\method{summary}{CureModel}(fit)
}
\arguments{
\item{formula}{An object of type \code{CureModel}}
}
\value{
An object of class \code{summary.CureModel}.
}
\description{
This function generates a summary of the fitted cure model.
}
