% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.survexp.dk.R
\docType{data}
\name{survexp.dk}
\alias{survexp.dk}
\title{Ratetable of the Danish general population}
\format{An object of class \code{ratetable} of dimension 111 x 180 x 2.}
\usage{
survexp.dk
}
\description{
Object of class \code{ratetable} containing the daily hazards in the Danish general population
as reported by the Human Mortality Database (www.mortality.org).
}
\details{
The ratetable was generated by using the \code{relsurv::transrate.hmd} function.
The data were downloaded on 15-09-2017 seperately for male and female Danish citizens.
The data can be accessed through:\cr
Female: http://www.mortality.org/hmd/DNK/STATS/fltper_1x1.txt\cr
Male: http://www.mortality.org/hmd/DNK/STATS/mltper_1x1.txt\cr
}
\keyword{datasets}
