% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threeClus.R
\docType{data}
\name{threeClus}
\alias{threeClus}
\title{Data set containing 3 spherical shaped clusters}
\format{
Columns x and y containing the x and y coordinates respectively, and a
column containing the true cluster label
}
\usage{
threeClus
}
\description{
Data set containing 3 spherical shaped clusters
}
\references{
https://github.com/Kchu/CURE-cluster-python - Data set named 3clus.txt here
}
\keyword{datasets}
