% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.R
\name{AIC.cm}
\alias{AIC.cm}
\title{Akaike information criterion}
\usage{
\method{AIC}{cm}(object, k = 2)
}
\arguments{
\item{object}{Object of class cm or gfcm.}

\item{k}{Number to control if either AIC or BIC is to be computed (default is 2 equal to AIC)}
}
\description{
This function computes the AIC for a fitted parametric cure model
}
