% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrudeProb.R
\name{calc.Crude}
\alias{calc.Crude}
\title{Crude event probabilities}
\usage{
calc.Crude(fit, newdata = NULL, type = "cancer", time = NULL,
  last.point = 100, reverse = FALSE, ci = T, expected = NULL,
  ratetable = survexp.dk, rmap = NULL, link = "logit")
}
\arguments{
\item{fit}{Fitted model to do predictions from. Possible classes are
\code{fmc}, \code{cm}, \code{stpm2}, and \code{pstpm2}.}

\item{newdata}{Data frame from which to compute predictions. If empty, predictions are made on the the data which
the model was fitted on.}

\item{type}{Measure to compute. Possible values are \code{cancer} (default),
\code{other}, and \code{othertime} (see details).}

\item{time}{Optional time points at which to compute predictions. If empty, a grid of 100 time points between 0
and \code{last.point} is selected.}

\item{last.point}{Constant at which the bound to tie probability is calculated. Default is 100.}

\item{reverse}{Logical. If \code{TRUE}, 1 - probability is provided (default is \code{FALSE}).
Only applicable for \code{type = othertime}.}

\item{ci}{Logical. If \code{TRUE} (default), confidence intervals are computed.}

\item{expected}{Object of class \code{list} containing objects of class \code{survexp},
with the expected survival of each row in newdata. If not specified, the function computes the expected
survival.}

\item{ratetable}{Object of class \code{ratetable} used to compute the general population survival.
Default is \code{survexp.dk}.}

\item{rmap}{List to be passed to \code{survexp} from the \code{survival} package.
Detailed documentation on this argument can be found by \code{?survexp}.}

\item{Link}{Link function for computing variance in order to bound confidence intervals. Default is \code{logit}.}
}
\value{
An object of class \code{crude} containing the crude probability estimates
of each individual in \code{newdata}.
}
\description{
Function for computing crude event probabilties from relative survival models.
}
\details{
The types of crude probabilities is the typical measures in relative survival, namely
a crude disease-specific probability (\code{cancer}), the probability of dying from other causes than the
disease (\code{other}), and the conditional probability of eventually dying from other causes than the
disease given survival until time t (\code{othertime}). The proportion of patients bound to die from the
disease can be computed by using \code{cancer} and choosing a sufficiently large time point.
}
\examples{
##Use data cleaned version of the colon cancer data from the rstpm2 package
data("colonDC")

##Extract general population hazards
colonDC$bhaz <- general.haz(time = "FU", age = "agedays", sex = "sex", year = "dx",
                            data = colonDC, ratetable = survexp.dk)

#Fit flexible parametric cure model
fit <- stpm2(Surv(FUyear, status2) ~ 1, data = colonDC, df = 6, bhazard = colonDC$bhaz, cure = T)

#Compute the probability of cancer related death
res <- calc.Crude(fit, time = seq(0, 20, length.out = 100),
                  rmap = list(age = agedays, sex = sex, year = dx))
plot(res)

#Compute the probability of eventually dying from other causes than cancer
res <- calc.Crude(fit, time = seq(0, 20, length.out = 100), type = "othertime",
                  rmap = list(age = agedays, sex = sex, year = dx))
plot(res)


#Fit parametric cure model
fit <- fit.cure.model(Surv(FUyear, status2) ~ 1, data = colonDC, bhazard = "bhaz",
                      formula.k1 = ~ 1, formula.k2 = ~ 1,
                      type = "mixture", dist = "weibull", link = "logit")

#Compute the probability of cancer related death
res <- calc.Crude(fit, time = seq(0, 20, length.out = 100),
                  rmap = list(age = agedays, sex = sex, year = dx))
plot(res)

#Compute the probability of eventually dying from other causes than cancer
res <- calc.Crude(fit, time = seq(0, 20, length.out = 100), type = "othertime",
                  rmap = list(age = agedays, sex = sex, year = dx))
plot(res)


}
\references{
Eloranta, S., et al. (2014) The application of cure models in the presence of competing risks: a tool
for improved risk communication in population-based cancer patient survival. \emph{Epidemiology}, 12:86.
}
