% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.LL.R
\name{calc.LL}
\alias{calc.LL}
\title{Loss of lifetime estimation}
\usage{
calc.LL(fit, newdata = NULL, time = NULL, tau = 100, ci = T,
  ratetable = survexp.dk)
}
\arguments{
\item{fit}{Fitted model to do predictions from. Possible classes are \code{fmcm}, \code{stpm2}, \code{pstpm2}, and \code{CureModel}.}

\item{newdata}{Data frame from which to compute predictions. If empty, predictions are made on the the data which
the model was fitted on.}

\item{time}{Optional time points at which to compute predictions. If empty, a grid of 100 time points between 0
and \code{tau} is selected.}

\item{tau}{The upper limit of the integral. Default is 100.}

\item{ci}{Logical indicating whether confidence intervals should be computed}

\item{ratetable}{Object of class \code{ratetable} to compute the general population survival from.}
}
\value{
A object of class \code{lol} containing the loss of lifetime estiamtes
of each individual in \code{newdata}.
}
\description{
Function for computing loss of lifetime estimates from an estimated relative survival model
}
