% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cure.R
\name{cure}
\alias{cure}
\title{Cluster two-dimensional data}
\usage{
cure(x, y, k, c, alpha, verbose = FALSE)
}
\arguments{
\item{x}{A vector of x-coordinates}

\item{y}{a vector of y-coordinates}

\item{k}{the number of desired clusters}

\item{c}{the number of representative points to use in each cluster}

\item{alpha}{the shrinkage factor to use}

\item{verbose}{Print remaining number of clusters? If TRUE, when number of clusters
is divisible by 10, the number of clusters is printed. Default is FALSE.}
}
\value{
A data frame containing columns for x, y, and the assigned cluster label
}
\description{
Cluster two-dimensional data
}
\examples{
cure(threeClus$x, threeClus$y, 3, 2, 0.25)

}
