% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.cure.model.R
\name{fit.cure.model}
\alias{fit.cure.model}
\title{Parametric cure model}
\usage{
fit.cure.model(formula, data, bhazard, formula.k1 = ~1, formula.k2 = NULL,
  formula.k3 = NULL, type = "mixture", dist = "weibull", link = "logit",
  covariance = TRUE, optim.args = NULL)
}
\arguments{
\item{formula}{An object of type formula for the cure rate.}

\item{data}{The provided data. Only objects of class \code{data.frame} is allowed.}

\item{bhazard}{The expected survival rate at the last follow-up of the individual.}

\item{formula.k1}{Formula for the first parameter of the parametric distribution (see details).}

\item{formula.k2}{Formula for the second parameter of the parametric distribution (see details).}

\item{formula.k3}{Formula for the third parameter of the parametric distribution (see details).}

\item{type}{Determines the type of cure model. Possible values are \code{mixture} (default) and \code{nmixture}.}

\item{dist}{The parametric distribution of the disease-specific survival function. Possible values are\cr
\code{weibull} (default), \code{exponential}, and \code{lognormal}.}

\item{link}{Specifies the link function of the cure rate. Possible values are \code{logistic} (default), \code{identity}, and \code{loglog}.}

\item{covariance}{Logical variable determining whether to compute the hessian or not. Default is \code{TRUE}}

\item{optim.args}{List with additional arguments to pass to \code{optim}.}

\item{init}{Initial values for the maximum likelihood optimization}
}
\value{
An object of class \code{CureModel}.
}
\description{
This function is used to fit parametric cure models on the relative survival.
}
\details{
The arguments for modelling the parameters of the cure model have different meanings dependent on the chosen distribution. \cr
For the exponential distribution, k1 denotes the rate, for the weibull model, k1 denotes the scale parameter and k2 denotes the shape parameter, sfor the log normal distribution k1 denotes the mu and sigma.
}
\examples{
##Use data cleaned version of the colon cancer data from the rstpm2 package
data("colonDC")

##Extract general population hazards
colonDC$bhaz <- general.haz(time = "FU", age = "agedays", sex = "sex", year = "dx",
                            data = colonDC, ratetable = survexp.dk)


####Without covariates
##Fit weibull cure model
fit <- fit.cure.model(Surv(FUyear, status2) ~ 1, data = colonDC, bhazard = "bhaz",
                      formula.k1 = ~ 1, formula.k2 = ~ 1,
                      type = "mixture", dist = "weibull", link = "logit")

##Plot model
plot(fit)
plot(fit, time = seq(0, 40, length.out = 100))
plot(fit, type = "ehaz")
plot(fit, type = "survuncured")
plot(fit, type = "probcure")


####With covariates
##Fit weibull cure model
fit <- fit.cure.model(Surv(FUyear, status2) ~ age, data = colonDC, bhazard = "bhaz",
                      formula.k1 = ~ age, formula.k2 = ~ 1,
                      type = "mixture", dist = "weibull", link = "logit")

##Plot model
plot(fit, newdata = data.frame(age = 50),
     time = seq(0, 15, length.out = 100), ci = F)
plot(fit, newdata = data.frame(age = 60),
     time = seq(0, 15, length.out = 100), col = 2, ci = F, add = T)

plot(fit, newdata = data.frame(age = 50),
     time = seq(0, 15, length.out = 100), ci = F, type = "ehaz")
plot(fit, newdata = data.frame(age = 60),
     time = seq(0, 15, length.out = 100), col = 2, ci = F, add = T, type = "ehaz")
}
