% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexMixtureModel.R
\name{FlexMixtureCureModel}
\alias{FlexMixtureCureModel}
\title{Fit spline-based mixture cure model}
\usage{
FlexMixtureCureModel(formula, data, bhazard, smooth.formula = ~1,
  knots = NULL, n.knots = NULL, tvc.formula = NULL, knots.time = NULL,
  n.knots.time = NULL, covariance = T, message = T, type = "mixture",
  linkpi = "logistic", linksu = "loglog")
}
\arguments{
\item{formula}{Formula for the covariates in the modelling of the cure rate, \eqn{\pi}. Reponse has to be of the form \code{Surv(time, status)}.}

\item{data}{Data frame in which to interpret the variables names in \code{formula}, \code{smooth.formula}, and \code{tvc.formula}.}

\item{bhazard}{Background hazard}

\item{smooth.formula}{Formula to control the modelling of the disease-specific survival of the uncured.}

\item{knots}{Knots used for the baseline hazard in the disease-specific survival function}

\item{n.knots}{Number of knots for the disease-specific survival function. The knots are calculated as the equidistant quantiles of the uncensored event-times.
If \code{knots} are supplied, this argument will be ignored.}

\item{tvc.formula}{Formula for the time-varying effects.}

\item{knots.time}{An object of class list containing the knots for each of the covariates in the time-varying covariate effects}

\item{n.knots.time}{An object of class list, containing the number of knots for the time-varying covariate effect.
If \code{knots.time} is supplied, this argument will be ignored.}

\item{type}{Character indicating which type of model is fitting.
Possible values are \code{mixture} (default) and \code{nmixture}.}

\item{linkpi}{Character for the type of link function selected for the cure rate.
Possible values are \code{logistic} (default), \code{identity}, \code{loglog}, and \code{probit}.}

\item{linksu}{Character for the type of link function selected for the survival of the uncured.
Possible values are \code{loglog}, \code{logistic}, and \code{probit}.}

\item{hes}{Logical for computing the inverse hessian matrix (default is \code{TRUE}).}

\item{verbose}{Logical indicating whether to output messages from the function}
}
\value{
An object of class \code{fmcm}.
}
\description{
The following function fits a generalized mixture or non-mixture cure model using a link function for the cure rate and for the survival of the uncured
}
