% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.lts.R
\name{lts}
\alias{lts}
\title{Long term survival predictions}
\usage{
lts(fit, newdata = NULL, time = NULL, ci = T, expected = NULL,
  ratetable = survexp.dk, rmap = rmap)
}
\arguments{
\item{fit}{Fitted model to do predictions from. Possible classes are \code{fcm}, \code{stpm2},
\code{pstpm2}, and \code{CureModel}.}

\item{newdata}{Data frame from which to compute predictions. If empty, predictions are made on the the data which
the model was fitted on.}

\item{time}{Optional time points at which to compute predictions. If empty, a grid of 100 time points between 0
and the maximum follow-up time is selected.}

\item{ci}{Logical indicating whether confidence intervals should be computed.}

\item{expected}{Object of class \code{list} containing objects of class \code{survexp}
denoting the expected survival of each row in newdata. If not specified, the function computes the expected
survival.}

\item{ratetable}{Object of class \code{ratetable} to compute the general population survival from.}

\item{rmap}{List to be passed to \code{survexp} from the \code{survival} package.
Detailed documentation on this argument can be found by \code{?survexp}.}
}
\value{
A object of class \code{lts} containing the loss of lifetime estiamtes
of each individual in \code{newdata}.
}
\description{
Function for computing loss of lifetime estimates from an estimated relative survival model
}
\examples{
##Use data cleaned version of the colon cancer data from the rstpm2 package
data("colonDC")

##Extract general population hazards
colonDC$bhaz <- general.haz(time = "FU", age = "agedays", sex = "sex", year = "dx",
                            data = colonDC, ratetable = survexp.dk)

##Fit flexible parametric relative survival model
fit <- stpm2(Surv(FUyear, status2) ~ 1, data = colonDC, df = 6, bhazard = colonDC$bhaz)

##Compute survival probabilities from 0 to 20 years
pred <- predict.lts(fit, time = seq(0, 20, length.out = 100),
                    rmap = list(age = agedays, sex = sex, year = dx))

##Plot the survival function
plot(pred)

##Fit flexible parametric cure model
fit <- FlexMixtureCureModel(Surv(FUyear, status2) ~ 1, data = colonDC, n.knots = 7, bhazard = "bhaz",
                            ini.types = "cure")

##Compute survival probabilities from 0 to 20 years
pred.cure <- predict.lts(fit, time = seq(0, 20, length.out = 100),
                         rmap = list(age = agedays, sex = sex, year = dx))

##Plot the loss of lifetime
plot(pred, ci = F)
plot(pred.cure, add = T, col = 2, ci = F)


}
