% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.crude.R
\name{plot.crude}
\alias{plot.crude}
\title{Plot crude event probabilities}
\usage{
plot.crude(obj, ylim = c(0, 1), xlim = NULL, ci = T, col = 1,
  ylab = NULL, xlab = "Time", add = F, ...)
}
\arguments{
\item{obj}{Object of class \code{crude} in which crude probabilities are calculated.}

\item{ylim}{Limits of y-axis.}

\item{xlim}{Limits of x-axis.}

\item{ci}{Logical denoting whether confidence intervals should be plotted (default is \code{TRUE}).}

\item{col}{Colour of each curve.}

\item{ylab}{Label of the y-axis. If \code{NULL}, the function uses its default labels depending on \code{obj$type}.}

\item{xlab}{Label of the x-axis (default is "Time").}

\item{add}{Logical indicating wether the cruves should be added to the current plot window (default is \code{FALSE}).}

\item{...}{Further arguments passed to \code{plot} and \code{lines}.}
}
\description{
Plot function for the computed crude event probabilties.
}
