% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.crude.R
\name{plot.crude}
\alias{plot.crude}
\title{Plot crude event probabilities}
\usage{
\method{plot}{crude}(obj, ylim = c(0, 1), xlim = NULL, ci = T, col = 1,
  ylab = NULL, xlab = "Time", add = F)
}
\arguments{
\item{obj}{Object of class \code{crude} in which crude probabilities are calculated}
}
\description{
Plot function for the computed crude event probabilties.
}
