% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.FlexCureModel.R
\name{plot.cuRe}
\alias{plot.cuRe}
\title{Plot function for Flexible mixture cure model}
\usage{
plot.cuRe(fit, newdata = NULL, type = "relsurv", time = NULL,
  ylim = c(0, 1), xlim = NULL, xlab = "Time", ylab = NULL,
  non.parametric = F, col = 1, col.non.para = 2, ci = T,
  include.knots = F, add = F, ...)
}
\arguments{
\item{fit}{Object of class \code{fmcm} to do predictions from.}

\item{newdata}{Data frame from which to compute predictions. If empty, predictions are made on the the data which
the model was fitted on.}

\item{type}{Type of prediction to do. Possible values are \code{relsurv} (default) for the relative survival,
\code{curerate} for the cure rate, \code{ehaz} for the excess hazard, and \code{probcure} for the
conditional probability of being cured.}

\item{time}{Optional time points at which to compute predictions. This argument is not used if type is \code{curerate}.}

\item{ci}{Logical indicating whether confidence intervals should be computed}

\item{pars}{Numerical vector containing the parameters values of the model.
In general, this argument can be ignored by the user}
}
\value{
A list containing the predictions of each individual in \code{newdata}.
}
\description{
Plot function associated with the class \code{fmcm}
}
