% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.le.R
\name{plot.le}
\alias{plot.le}
\title{Plot the loss of lifetime function}
\usage{
\method{plot}{le}(x, ylim = NULL, xlim = NULL, ci = T, col = 1,
        ylab = NULL, xlab = "Time", add = F, \dots)
}
\arguments{
\item{x}{Object of class \code{le}.}

\item{ylim}{Limit of the y-axis.}

\item{xlim}{Limit of x-axis.}

\item{ci}{Logical. If \code{TRUE} (default), confidence intervals are added to the plot.}

\item{col}{Numeric or character indicating the colours of the curves.}

\item{ylab}{Label to be written on the y-axis.}

\item{xlab}{Label to be written on the x-axis.}

\item{add}{Logical. If \code{TRUE}, the curve is added to the current plot window.}

\item{...}{Further argument passed to \code{plot} and \code{lines}.}
}
\description{
Plot function for the class \code{le}
}
