% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.lts.R
\name{plot.lts}
\alias{plot.lts}
\title{Plot the long term survival}
\usage{
\method{plot}{lts}(obj, ylim = NULL, xlim = NULL, ci = T, col = 1,
  ylab = NULL, xlab = "Time", add = F, ...)
}
\arguments{
\item{obj}{Object of class \code{lts}.}

\item{ylim}{Limit of the y-axis. If empty, these are chosen occording to the extremes in \code{obj}.}

\item{xlim}{Limit of x-axis. If empty, these are chosen according to the time extremes of \code{obj}.}

\item{ci}{Logical indicating wether confidence intervals should be included in the plot (default is TRUE).}

\item{col}{Numeric or character indicating the colours of the curves.}

\item{ylab}{Label to be written on the y-axis.}

\item{xlab}{Label to be written on the x-axis.}

\item{add}{Logical indicating whether to add to current plot window (default is FALSE).}

\item{...}{Further argument passed to \code{plot}.}
}
\description{
Plot function for the class \code{lts}
}
