% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.cm.R
\name{predict.cm}
\alias{predict.cm}
\title{Predict function for cure models}
\usage{
predict.cm(fit, newdata = NULL, type = "relsurv", time = NULL, ci = T,
  pars = NULL)
}
\arguments{
\item{fit}{An object of type CureModel.}

\item{newdata}{A dataframe with new data to predict from.}

\item{type}{A character denoting the type of prediction. Possible values are \code{relsurv} (default), \code{curerate}, and \code{probcure} (see details)}

\item{ci}{Logical indicating whether to include confidence intervals. Default is \code{TRUE}}

\item{times}{A numeric vector including the time points for which to make predictions.}
}
\value{
An object of class \code{matrix} including the predictions.
}
\description{
This function is used to make predictions of the cure models.
}
\details{
For type \code{relsurv}, relative survival predictions are made, while for \code{curerate}, \eqn{\pi} and for \code{probcure}\cr
the probability of cure is calculated.
}
