% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_pts.R
\name{rep_pts}
\alias{rep_pts}
\title{Find the representative points of a cluster}
\usage{
rep_pts(x, npt = 3, alpha = 0.5)
}
\arguments{
\item{x}{A data frame containing the points in the cluster}

\item{npt}{Number of representative points to use}

\item{alpha}{Shrinkage parameter}
}
\value{
A data frame containing the x and y coordinates of the representative points
}
\description{
Find the representative points of a cluster
}
\examples{
clus_pts <- data.frame(x = iris$Sepal.Length, y = iris$Sepal.Width, clus.lab = 1:length(iris$Sepal.Width))
rep_pts(clus_pts, 3, 0.75)
}
